move xmlAns to it own class
